fix fetch and models
